Helm no longer auto-detects the version, it is supplied by the user
