+++ source/Calamari.Shared/Util/HelmVersion.cs
@@ -1,0 +1,8 @@
+namespace Calamari.Util
+{
+    public enum HelmVersion
+    {
+        V2,
+        V3
+    }
+}
