+++ source/Calamari/Kubernetes/SpecialVariables.cs
@@ -18,6 +18,7 @@
             public const string ResetValues = "Octopus.Action.Helm.ResetValues";
             public const string AdditionalArguments = "Octopus.Action.Helm.AdditionalArgs";
             public const string CustomHelmExecutable = "Octopus.Action.Helm.CustomHelmExecutable";
+            public const string ClientVersion = "Octopus.Action.Helm.ClientVersion";
             public const string Timeout = "Octopus.Action.Helm.Timeout";
             public const string TillerNamespace = "Octopus.Action.Helm.TillerNamespace";
             public const string TillerTimeout = "Octopus.Action.Helm.TillerTimeout";
